Discord Bot/requirements.txt
aiohttp==3.7.3
apturl==0.5.2
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==21.2.1
Automat==20.2.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dateparser==1.0.0
dbus-python==1.2.16
defer==1.0.6
discord.py==1.6.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
docker==4.1.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
isort==5.7.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
multidict==5.1.0
netifaces==0.10.4
numpy==1.20.1
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
pika==0.11.0
Pillow==7.0.0
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.4
python-binance==0.7.9
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2020.11.13
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
systemd-python==234
texttable==1.6.2
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==4.0.2
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
websocket-client==0.53.0
wrapt==1.12.1
xkit==0.0.0
yarl==1.6.3
zipp==1.0.0
zope.interface==5.2.0
